third_party/blink/web_tests/external/wpt/css/css-text/parsing/text-transform-computed-expected.txt

This is a testharness.js-based test.
PASS Property text-transform value 'none' computes to 'none'
PASS Property text-transform value 'capitalize' computes to 'capitalize'
PASS Property text-transform value 'uppercase' computes to 'uppercase'
PASS Property text-transform value 'lowercase' computes to 'lowercase'
FAIL Property text-transform value 'full-width' computes to 'full-width' assert_equals: expected "full-width" but got "none"
FAIL Property text-transform value 'full-size-kana' computes to 'full-size-kana' assert_equals: expected "full-size-kana" but got "none"
FAIL Property text-transform value 'capitalize full-width' computes to 'capitalize full-width' assert_equals: expected "capitalize full-width" but got "none"
FAIL Property text-transform value 'full-width full-size-kana' computes to 'full-width full-size-kana' assert_equals: expected "full-width full-size-kana" but got "none"
FAIL Property text-transform value 'uppercase full-width full-size-kana' computes to 'uppercase full-width full-size-kana' assert_equals: expected "uppercase full-width full-size-kana" but got "none"
Harness: the test ran to completion.
